# docs/openapi/reconcile-v1.yaml
openapi: 3.1.0
info:
  title: Reconcile - payment & settlement reconciliation
  description: |
    Money movement, provider settlement ingestion, and reconciliation.

    Every mutating endpoint requires an `Idempotency-Key` and replays a
    stored response for 24 hours. Refusals are RFC 9457 problem
    documents; the failure taxonomy is in `docs/spec/07-failure-matrix.md`
    and the error table in `docs/spec/06-api-contract.md`.

    Two routes are deliberately unauthenticated: `GET /api/v1/health`,
    which a load balancer must be able to reach, and
    `POST /api/v1/provider/webhooks`, which authenticates by HMAC
    signature rather than by token.
  license:
    name: Proprietary
  version: v1
servers:
- url: /
  description: This application
security:
- bearerAuth: []
paths:
...
  /api/v1/provider/webhooks:
    post:
      tags:
      - provider-webhook-controller
      operationId: receive
      parameters:
      - name: X-Provider-Event-Id
        in: header
        required: false
        schema:
          type: string
      - name: X-Provider-Timestamp
        in: header
        required: false
        schema:
          type: string
      - name: X-Provider-Signature
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |
        A static operator or admin bearer token. Admin implies
        operator. Not a production identity design; see the
        authentication section of the README for what a real
        deployment puts in front of this instead.
      scheme: bearer